[TEX-537] Optional support for unframing `grpc-web+json` requests. (#501)
Add a new option for unframing `application/grpc-web+json` requests to
be just `application/json` requests if the request is to a whitelisted
hostname. This unframing is performed by mutating the received request
before doing anything else with it, including passing it upstream or
searching for existing tapes. This has the effect that, for all intents
and purposes, the request is an `application/json` request for the
purposes of persistence, similarity scoring, and proxying.