[TEX-537] Refactor HTTP request and response representations (#554)
This PR does a bunch of initial refactoring and cleaning up of the
representations of HTTP request and response objects throughout the
codebase, ahead of making some larger changes to support gRPC-web
request comparisons.

The changes in this PR are mostly code structure in nature rather than
functional. There is one functionality change in the decision on how to
compare two HTTP request bodies, which I've called out in comments.