[TEX-537] Add support for gRPC-web request body inspection (#548)
Adds support to proxay to perform gRPC-web request body inspection when computing request similarity. This is done by inspecting the content type of the request. If it's `application/grpc-web` or `application/grpc-web-text`, we invoke new body-to-object decoding logic, like we do for `application/json` requests. If this body-to-object decoding logic returns an object, the normal object similarity scoring happens, which is what we're after all along.