Ignore Headers / Send Proxy Port / Dump Matcher Fails (#604)
I added three flags that are needed in my situation and might be helpful
for others:

`--ignore-headers`: Allows users to specify a list of headers that
should be ignored by Proxay. This is especially useful together with the
`--exact-request-matching` flag.

`--send-proxy-port`: Sends the proxays port to the proxied host, similar
to nginx' `proxy_set_header Host $host:$server_port;`. This helps
against redirect issues.

`--dump-matcher-fails`: Dumps the headers from the current and recorded
request, if they are not matching. This helps to find headers that
should be ignored in the specific situation.

For all three parts there are simple tests. I also added a paragraph in
the Readme that documents those flags.

---------

Co-authored-by: Tim Dawborn <tim.dawborn@gmail.com>